+++ client/src/response.ts
@@ -1,7 +1,6 @@
 
 import { ViewId } from './action'
-import { takeWhileMap } from "./lib"
-import { actionMessage, ActionMessage, Metadata } from "./action"
+import { Metadata } from "./action"