+++ src/Web/Hyperbole/Page.hs
@@ -2,9 +2,9 @@ module Web.Hyperbole.Page where
 
 import Data.Kind (Type)
 import Effectful
-import Web.Hyperbole.Effect.Handler (RunHandlers, runLoad)
 import Web.Hyperbole.Effect.Hyperbole
 import Web.Hyperbole.HyperView (Root)
+import Web.Hyperbole.Server.Handler (RunHandlers, runLoad)
 import Web.Hyperbole.Types.Response (Response)
 import Web.Hyperbole.View (View)